Add playbook for the 1.5 Jenkins
And an environment variable we can use to select this Jenkins. This will allow us to start using Ansible to manage that Jenkins.

The name is chosen to discourage use - it's currently not safe to run Ansible against the 1.5 Jenkins unless you're very confident in what Ansible is doing and how the 1.5 Jenkins is configured. There are some Ansible tasks (e.g. jobs) that would actively destroy the 1.5 Jenkins if they were run.