Manually source ~/.profile to enable rbenv
In the latest version (2.332.1), Jenkins moved to systemd. I still don't understand why, but this has had an impact on our builds. Previously, the shell used in builds would source ~/.profile, which is needed for rbenv to work. With the new version, this doesn't happen any more.

I've not worked out the 'right' way to do this. So I think a vaguely acceptable alternative is to manually source the .profile file at the start of every step that uses Ruby. It's unsatisfactory, but I can't see any alternative.