Use the system Ruby for tests instead of rbenv
Since we upgraded to Jenkins 2.332.1, rbenv hasn't worked correctly. I'm not sure why, but Jenkins moved from SysV to systemd. As far as I can tell, the shells Jenkins creates for builds no longer source ~/.profile, which is needed for rbenv to work.

I don't think there's a better alternative. I can't work out how to get the tests working successfully in Docker - at least not in parallel mode. I recently added a bodge to source ~/.profile at the start of each job that needed it. This mostly worked, and allowed tests to resume. However, we're still getting failures, and I'm not sure why.

So I think the least worst option is to abandon the use of rbenv on Jenkins in favour of the system version (2.7.0p, at the moment). I think that's fine because we can still use rbenv locally where it matters most, and nothing else on Jenkins is using Ruby.