Create pipeline to rotate AWS tokens
This is a task we need to perform once a quarter for preview and production. I've enhanced rotate-api-tokens.sh to support rotating the AWS token, so use it in a pipeline. Should make the rotations easier and more straightforward.

Won't be able to test this without merging and deploying to Jenkins. But I copied liberally from the other rotation jobs, so I think it's likely to work.