Run job to create approved service edit report monthly
A category admin wants a monthly report of approved service edits. We've been triggering this manually, but this is getting a bit tedious. Add a Jenkins job to create the report, upload it to S3, and send a notification email.

I've tested this against preview, and it seems to work fine.